potentially making it worse
Messed with a few things, just added some random color to the footer to show I know how to mess with that in particular. I also forgot a semicolon somewhere so I added that, and took out some numbers I don't remember writing that didn't seem to actually serve a purpose.